cash_flow/requirements.txt
django>=5.2
djangorestframework